Split apart various target type code.
This splits apart the "generators" that fill the Target objects from script, and the "writers" that write the .ninja files. These now have different objects and files for the basic target types. The result is that it looks like much
more stuff due to the extra files, but it should be easier to follow what's happening and to make additions in the future.

Remove some extra loadable module stuff I missed in the previous removal.

Adds more support for the Linux build. Not everything is working yet, but most of base compiles.

This actually writes .ninja files and targets for "group" types, and removes the "none" target type which this mapped to before. This way you can make a target group and actually compile that set of stuff by typing its name.

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/22488015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@217279 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c0822d7f9288ed13dce228ae3c3ed080e9712e3f

+++ tools/gn/input_conversion.cc
@@ -36,7 +36,7 @@ std::string GetExampleOfBadInput(const std::string& input) {
   }
   TrimWhitespaceASCII(result, TRIM_ALL, &result);
 
-  const int kMaxSize = 50;
+  const size_t kMaxSize = 50;
   if (result.size() > kMaxSize) {
     trimmed = true;
     result.resize(kMaxSize);
